dataprocessing/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Data processing for the chatGPITpt project. The responsibilities of this subproject are downloading the git history, making the embeddings and scheduling this in a pipeline run."
authors = ["ChidiRNweke <chidi125@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
dagster = "^1.7.8"
dagster-webserver = "^1.7.8"
pygithub = "^2.3.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
module_name = "src"
